+++ test/e2e/advanced/conversation-encryption.3.spec.js
@@ -120,9 +120,7 @@ describe('Conversation encryption -', function () {
             it("check security aspects", function () {
                 util.waitForElement('cm-header:not(.ng-hide)')
 
-                ptor.debugger()
-
-                var p,n 
+                var p,n
 
                 ptor.wait(function(){
 
@@ -223,18 +221,22 @@ describe('Conversation encryption -', function () {
                             util.setVal('input-password', password)
 
                             ptor.wait(function(){
-                                return util.getVal('input-password').then(function(val){
+                                return util.getVal('input-password')
+                                .then(function(val){
                                     return val == password
                                 })
+                            }).then(function(){
+                                // make sure that the input loses focus and ng-blur gets fired:
+                                $("[data-qa='input-password']").sendKeys(protractor.Key.TAB)
+                                $("#cm-app").click()
+
+                                return util.waitForElement("[data-qa='icon-conversation-decrypted']")
+                                    .then(function(){
+                                        return $("[data-qa='btn-security-done']").click()
+                                    })
+                            }).then(function(){
+                                return util.waitForPageLoad(conversationRoute)
                             })
-
-                            //make sure that the input loses focus and ng-blur gets fired:
-                            $("[data-qa='input-password']").sendKeys(protractor.Key.TAB)
-                            $("#cm-app").click()
-
-                            util.waitForElement("[data-qa='icon-conversation-decrypted']")
-                            $("[data-qa='btn-security-done']").click()
-                            util.waitForElementDisappear("[data-qa='btn-security-done']")
                         }
                     }
                 })